backend/requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Security & Authentication
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==2.3.0

# API Documentation
drf-spectacular==0.26.5

# Task Queue
celery==5.3.4
redis==5.0.1

# File Processing
python-magic==0.4.27
Pillow==10.0.1

# Firmware Analysis
capstone==4.0.2
pefile==2023.2.7
pyelftools==0.29

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# Optional: Deep Learning
# tensorflow==2.13.0
# torch==2.0.1

# Network & Communication
requests==2.31.0
websockets==11.0.3

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.6.0
tqdm==4.66.1

# Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1

# Monitoring & Logging
sentry-sdk==1.32.0
structlog==23.1.0
